Fixed Mine Bug, Mining Node bug

+++ Calypso/Assets/Scripts/Mining/MiningInteractable.cs
@@ -8,7 +8,7 @@ public class MiningInteractable : MonoBehaviour, IInteractable
     [SerializeField] private ItemDrop resourceDrop;
     [SerializeField] private Transform spawnPoint;
     [SerializeField] private float mineTime = 2f;
-    [SerializeField] private int Maxhp = 3;
+    [SerializeField] private float Maxhp = 3;
     [SerializeField] private GameObject model;
 
     [Header("Particle Config")]
@@ -21,7 +21,7 @@ public class MiningInteractable : MonoBehaviour, IInteractable
     [Header("Event Object")]
     [SerializeField] private OnGenericEventSO clearPlayerInteractableEvent;
 
-    private int hp = 1;
+    private float hp = 1;
     private float currentMineTimer = 0f;
     private bool mining = false;
     private MiningParticleController particleController;
